requirements.txt
absl-py==1.0.0
asttokens==2.0.8
astunparse==1.6.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
brotlipy==0.7.0
cachetools==5.0.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///C:/ci/cffi_1636542074879/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
commonmark==0.9.1
conda==22.9.0
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262320430/work
cryptography @ file:///C:/ci/cryptography_1639472366776/work
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
docker-pycreds==0.4.0
entrypoints @ file:///C:/ci/entrypoints_1649926621247/work
executing==1.0.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
hexdump==3.3
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.10.1
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/b/abs_84shimxwmz/croot/ipython_1668088123494/work
jedi @ file:///C:/ci/jedi_1644315425835/work
joblib==1.1.0
jupyter_client @ file:///C:/b/abs_8fbm7986b_/croots/recipe/jupyter_client_1662504374117/work
jupyter_core @ file:///C:/b/abs_b8r5lb5eus/croot/jupyter_core_1668084445779/work
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.2
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934035815/work
menuinst==1.4.16
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929372/work
numpy==1.23.1
oauthlib==3.1.1
onnx==1.12.0
onnx-simplifier==0.4.0
onnxruntime==1.11.1
onnxsim-no-ort==0.4.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.1
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.12.0
PyJWT==2.3.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pyupbit==0.2.22
pywin32==302
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657616005830/work
requests==2.28.1
requests-oauthlib==1.3.0
rich==12.5.1
rsa==4.8
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016967756/work
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smmap==5.0.0
stack-data==0.5.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
thop==0.1.0.post2207010342
threadpoolctl==3.1.0
tinymlgen==0.2
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado @ file:///C:/ci/tornado_1662476991259/work
tqdm @ file:///C:/ci/tqdm_1650618053064/work
traitlets==5.4.0
typing_extensions==4.0.1
urllib3==1.26.7
wandb==0.12.21
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==9.1
Werkzeug==2.0.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.13.3
zipp==3.7.0
